Get sack from package, this allows converting package to a PackageSet
This is used in AdvisoryQuery so that one package can be converted to a
PackageSet with one package and this can be handled by the AdvisoryQuery
just like any other PackageSet.

It should be noted that I am not sure if we actually need the general
filter with PackageSet. Maybe filtering just by one package is enought
and we can remove this.